fix(direct): prevent double-badge on /direct/ icon for rooms in sidebar
The /direct/ icon badge was counting unread for all DMs, including
those already displayed as individual avatar entries in the sidebar
via DirectDMsList. This caused users to see duplicate badges — one
on the avatar entry and one on the /direct/ icon — for the same room.

Extract useSidebarDirectRoomIds hook (shared by DirectDMsList and
DirectTab) that returns the room IDs currently shown as sidebar
avatars. DirectTab now computes its badge only for DMs that overflow
beyond the MAX_SIDEBAR_DMS (3) limit shown in the sidebar.

Closes #235
